Refactor utils to core for algos
